+++ src/with-slices.ts
@@ -1,9 +1,5 @@
 import type { SliceConfig } from './create-slice.js';
 
-type ParametersIf<T> = T extends (...args: infer Args) => unknown
-  ? Args
-  : never;
-
 type InferState<Configs> = Configs extends [
   SliceConfig<infer Name, infer Value, infer Actions>,
   ...infer Rest,
@@ -15,37 +11,17 @@ type InferState<Configs> = Configs extends [
     } & InferState<Rest>
   : unknown;
 
-type HasDuplicatedNames<
-  Configs,
-  Names extends string[] = [],
-> = Configs extends [
+type HasDuplicatedNames<Configs, Names = never> = Configs extends [
   SliceConfig<infer Name, infer _Value, infer Actions>,
   ...infer Rest,
 ]
-  ? Extract<Name | keyof Actions, Names[number]> extends never
-    ? HasDuplicatedNames<Rest, [Name, ...Names]>
-    : true
-  : false;
-
-type HasDuplicatedArgs<Configs, State> = Configs extends [
-  SliceConfig<infer _Name, infer _Value, infer Actions>,
-  ...infer Rest,
-]
-  ? {
-      [actionName in keyof State]: ParametersIf<State[actionName]>;
-    } extends {
-      [actionName in keyof Actions]: Parameters<Actions[actionName]>;
-    }
-    ? HasDuplicatedArgs<Rest, State>
+  ? Extract<Name | keyof Actions, Names> extends never
+    ? HasDuplicatedNames<Rest, Name | keyof Actions | Names>
     : true
   : false;
 
 type ValidConfigs<Configs> =
-  HasDuplicatedNames<Configs> extends true
-    ? never
-    : HasDuplicatedArgs<Configs, InferState<Configs>> extends true
-      ? never
-      : Configs;
+  HasDuplicatedNames<Configs> extends true ? never : Configs;
 
 export function withSlices<
   Configs extends SliceConfig<string, unknown, NonNullable<unknown>>[],
@@ -63,29 +39,17 @@ export function withSlices<
   ) => {
     const state: Record<string, unknown> = {};
     type ActionFn = (...args: unknown[]) => (prev: unknown) => unknown;
-    const sliceMapsByAction = new Map<string, Map<string, ActionFn>>();
     for (const config of configs) {
       state[config.name] = config.value;
-      for (const [actionName, actionFn] of Object.entries(config.actions)) {
-        let actionsBySlice = sliceMapsByAction.get(actionName);
-        if (!actionsBySlice) {
-          sliceMapsByAction.set(actionName, (actionsBySlice = new Map()));
-        }
-        actionsBySlice.set(config.name, actionFn as never);
+      for (const [actionName, actionFn] of Object.entries<ActionFn>(
+        config.actions,
+      )) {
+        state[actionName] = (...args: unknown[]) => {
+          set(((prevState: Record<string, unknown>) => ({
+            [config.name]: actionFn(...args)(prevState[config.name]),
+          })) as never);
+        };
       }
-    }
-    for (const [actionName, actionsBySlice] of sliceMapsByAction) {
-      state[actionName] = (...args: unknown[]) => {
-        set(((prevState: Record<string, unknown>) => {
-          const nextState: Record<string, unknown> = {};
-          for (const [sliceName, actionFn] of actionsBySlice) {
-            const prevSlice = prevState[sliceName];
-            const nextSlice = actionFn(...args)(prevSlice);
-            nextState[sliceName] = nextSlice;
-          }
-          return nextState;
-        }) as never);
-      };
     }
     return state;
   }) as never;